Add Soundboard class, add sound to blockdude.py
This commit implements the first half of a soundboard which can be used
by game elements to add sounds to their actions. In the main code base,
only game.py and elements/element_data.py were edited to add soundboard
functionality and the elements/blockdude.py class implements it

Only SFX are implemented for now, music requires that certain channels be reserved
and a way to manage the loop. This commit also does not prevent key
mashing in the main block dude example. Stepping and bumping sounds are
added to the main blockdude.py example.

Additionally, this commit adds a whole bunch of new sfx to choose from
and moves them to the "src/sounds/sfx" directory.

Interactive test "src/sounds/test-soundboard.py" was added to the test
the Soundboard class specifically (in the future, should be merged with
the other test)

black formatting